// ecommerce-admin-master/src/app/model/product.create.title.model.ts
export class ProductCreateTitleModel {
    name: string;
    Information: string;
    Specifications: string;
    Images:string;
    Variants: string;
    Reviews: string;
    Name: string;
    Alias: string;
    Ifthealiasisexist: string;
    Type: string;
    Physical: string;
    Digital: string;
    DigitalFilePath: string;
    StockQuantity: string;
    Seller: string;
    searching:string;
    Sorrysuggestionscouldnotbeloaded:string;
    Price: string;
    SalePrice: string;
    TaxName: string;
    TaxValue:string;
    RestrictCODareas: string;
    Add: string;
    Shortdescription: string;
    Description: string;
    Category: string;
    Nocategory: string;
    Metakeywords: string;
    Metadescription: string;
    Enterzipcode: string;
    ZipCode: string;
    FreeShipCountry: string;
    FreeShipState: string;
    FreeShipCity: string;
    Dailydeal: string;
    DealTo:string;
    FreeShip: string;
    Featured: string;
    HotProduct: string;
    BestSelling: string;
    Active: string;
    isPromotion: string;
    Clicktosetmainimage: string;
    Pleasecreateproductthenaddvariants: string;
    Pleasecreateproduct: string;
    Submit:string;
    Cancel: string;
    Weight:string;
    TradeDiscount: string;
    WeightCondition: string;
    MarketingDiscount: string;
    BuyerDiscount: string;
    GoldTimeDiscount: string;
    Deny: string;
    Cause: string;
    SAP: string;
    NTDPrice: string;
    NTDUnit: string;
    WEPrice: string;
    WEUnit: string;
    Lang: string;
    SKU: string;
    PackageMethod: string;
    Producer: string;
    Expiry: string;
    Country: string;
    Ingredient: string;
}
 export const ProductCreateTitle: ProductCreateTitleModel[] = [
     {  
        name: '',
        Information: '',
        Specifications: '',
        Images:'',
        Variants: '',
        Reviews: '',
        Name: '',
        Alias:'',
        Ifthealiasisexist: '',
        Type: '',
        Physical: '',
        Digital: '',
        DigitalFilePath: '',
        StockQuantity: '',
        Seller: '',
        searching:'',
        Sorrysuggestionscouldnotbeloaded:'',
        Price:'',
        SalePrice: '',
        TaxName: '',
        TaxValue:'',
        RestrictCODareas: '',
        Add: '',
        Shortdescription: '',
        Description: '',
        Category: '',
        Nocategory: '',
        Metakeywords: '',
        Metadescription: '',
        Enterzipcode: '',
        ZipCode: '',
        FreeShipCountry: '',
        FreeShipState: '',
        FreeShipCity: '',
        Dailydeal: '',
        DealTo:'',
        FreeShip: '',
        Featured:'',
        HotProduct: '',
        BestSelling: '',
        Active: '',
        isPromotion: '',
        Clicktosetmainimage: '',
        Pleasecreateproductthenaddvariants:'',
        Pleasecreateproduct: '',
        Submit:'',
        Cancel: '',
        Weight:'',
        TradeDiscount:'',
        WeightCondition: '',
        MarketingDiscount: '',
        BuyerDiscount: '',
        GoldTimeDiscount: '',
        Deny: '',
        Cause: '',
        SAP: '',
        NTDPrice: '',
        NTDUnit: '',
        WEPrice: '',
        WEUnit: '',
        Lang: '',
        SKU: '',
        PackageMethod: '',
        Producer: '',
        Expiry: '',
        Country: '',
        Ingredient: ''
    },
 ]